Rename namespace Dwt.Api.Middleware.Jwt to Dwt.Api.Middleware.JwtIdentity
