+++ content/Dwt/Dwt.Api/Middleware/JwtIdentity/JwtIdentityServiceCollectionExtensions.cs
@@ -1,4 +1,4 @@
-﻿namespace Dwt.Api.Middleware.Jwt;
+﻿namespace Dwt.Api.Middleware.JwtIdentity;
 
 public static class JwtIdentityServiceCollectionExtensions
 {
